feat: enhance command success listeners with audit logging for chat, context, and message commands
