+++ src/listeners/commands/chatInputCommandSuccessLogger.ts
@@ -1,6 +1,5 @@
-import { readSettingsAuditLog, readSettingsCached } from '#lib/database';
 import { ApplyOptions } from '@sapphire/decorators';
-import { Events, Listener, type ChatInputCommandSuccessPayload } from '@sapphire/framework';
+import { Events, Listener, LogLevel, type ChatInputCommandSuccessPayload } from '@sapphire/framework';
 import { cyan } from 'colorette';
 import type { User } from 'discord.js';
 
@@ -12,17 +11,11 @@ export class UserListener extends Listener<typeof Events.ChatInputCommandSuccess
 		const author = this.author(interaction.user);
 		const sentAt = interaction.guildId ? `${interaction.guild?.name ?? 'Unknown'}[${cyan(interaction.guildId)}]` : cyan('Direct Messages');
 		this.container.logger.debug(`${shard} - ${commandName} ${author} ${sentAt}`);
+	}
 
-		if (!interaction.guildId) return;
-		const settings = readSettingsCached(interaction.guildId);
-		if (!settings) return;
-		void readSettingsAuditLog(settings)
-			.command(interaction.user.id, {
-				commandName: interaction.commandName,
-				commandType: 'chat-input',
-				channelId: interaction.channelId
-			})
-			.catch(() => null);
+	public override onLoad() {
+		this.enabled = this.container.logger.has(LogLevel.Debug);
+		return super.onLoad();
 	}
 
 	private author(author: User) {
