+++ src/listeners/commands/messageCommandSuccessLogger.ts
@@ -1,6 +1,5 @@
-import { readSettingsAuditLog, readSettingsCached } from '#lib/database';
 import { ApplyOptions } from '@sapphire/decorators';
-import { Command, Events, Listener, type MessageCommandSuccessPayload } from '@sapphire/framework';
+import { Command, Events, Listener, LogLevel, type MessageCommandSuccessPayload } from '@sapphire/framework';
 import { cyan } from 'colorette';
 import type { Guild, User } from 'discord.js';
 
@@ -12,17 +11,11 @@ export class UserListener extends Listener<typeof Events.MessageCommandSuccess> 
 		const author = this.author(message.author);
 		const sentAt = message.guild ? this.guild(message.guild) : this.direct();
 		this.container.logger.debug(`${shard} - ${commandName} ${author} ${sentAt}`);
+	}
 
-		if (!message.guildId) return;
-		const settings = readSettingsCached(message.guildId);
-		if (!settings) return;
-		void readSettingsAuditLog(settings)
-			.command(message.author.id, {
-				commandName: command.name,
-				commandType: 'message',
-				channelId: message.channelId
-			})
-			.catch(() => null);
+	public override onLoad() {
+		this.enabled = this.container.logger.has(LogLevel.Debug);
+		return super.onLoad();
 	}
 
 	private shard(id: number) {
